feat: write several backend API methods + fix integration with docker compose

+++ data/schema.sql
@@ -7,8 +7,8 @@ CREATE TABLE Patterns (
 CREATE TABLE Effects (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     breathing REAL NOT NULL,
-    chasing INTEGER NOT NULL,
-    sparkle REAL NOT NULL,
+    chasing REAL NOT NULL,
+    sparkle INTEGER NOT NULL,
     FOREIGN KEY (id) REFERENCES Patterns(id)
 );
